test(drivers): small reorganization of some tests
